Add "n/a" to sliders as default value
Instead of 1 (min) being the default value, set it to 0 ("None"). Moving the slider to other values selects that value, moving it back to n/a clears it.

May need updating in the future. Will only work where 0 is not a valid value. Might be problematic if "min" is greater than 1? But will do for neuroaesthetic scores