WIP: Just doing this doesn't work (#1369)
If you set this field to Required, then it is required for all fields in all subcategories, not just the current subcategory.

We need to find a way to restrict this, so the field is only required when certain fields are changed, without resorting to writing lots of extra code (like in the Community category).